refactor: extrait useReactTable du composant Tableau (318)
